+++ src/main/java/pe/edu/colegiocima/appcima/service/GradoService.java
@@ -6,8 +6,8 @@ import pe.edu.colegiocima.appcima.models.entity.Grado;
 
 public interface GradoService {
     public Iterable<Grado> findAll();
-    public void deleteById(Long id);
-    public Grado findById(Long id);
+    public void deleteById(Short id);
+    public Grado findById(Short id);
     public Grado save(Grado grado);
     public Page<Grado> findByIdNivelColegioOrderByDescripcion(Integer idNivelColegio, Pageable pageable);
 }
